Server/requirements.txt
Flask>=3.1.2
Flask-Cors>=6.0.1
openai>=2.9.0
python-dotenv>=1.2.1
tiktoken>=0.12.0
